Changed floating-point precision

+++ market.h
@@ -1,7 +1,7 @@
 #include "list.h"
 struct s_luld
 {
-    int min, max;
+    double min, max;
 };
 typedef struct s_luld t_luld;
 
@@ -14,7 +14,7 @@ typedef struct s_order t_order;
 
 struct s_ndist
 {
-    float mean;
-    float variance;
+    double mean;
+    double variance;
 };
 typedef struct s_ndist t_ndist;
